requirements.txt
pytest==6.0.1
pytest-pycodestyle==2.2.0
requests
flask
testfixtures
pymodm
dnspython
numpy
pillow
pandas
scipy
matplotlib
